Merge Dev-2

+++ server/DocumentsKM/Dtos/ConstructionBolt/ConstructionBoltResponse.cs
@@ -1,14 +1,15 @@
+using System;
 using DocumentsKM.Models;
 
 namespace DocumentsKM.Dtos
 {
     public class ConstructionBoltResponse
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
         public BoltDiameter Diameter { get; set; }
-        public int Packet { get; set; }
-        public int Num { get; set; }
-        public int NutNum { get; set; }
-        public int WasherNum { get; set; }
+        public Int16 Packet { get; set; }
+        public Int16 Num { get; set; }
+        public Int16 NutNum { get; set; }
+        public Int16 WasherNum { get; set; }
     }
 }
