+++ server/DocumentsKM/Dtos/GeneralDataPoint/GeneralDataPointResponse.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class GeneralDataPointResponse
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
         public string Text { get; set; }
-        public int OrderNum { get; set; }
+        public Int16 OrderNum { get; set; }
     }
 }
