+++ server/DocumentsKM/Dtos/GeneralDataPoint/GeneralDataPointUpdateRequest.cs
@@ -1,9 +1,11 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class GeneralDataPointUpdateRequest
     {
         public string Text { get; set; }
-        public int? OrderNum { get; set; }
+        public Int16? OrderNum { get; set; }
 
         public GeneralDataPointUpdateRequest()
         {
