+++ server/DocumentsKM/Dtos/MarkLinkedDoc/MarkLinkedDocCreateRequest.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace DocumentsKM.Dtos
@@ -5,7 +6,7 @@ namespace DocumentsKM.Dtos
     public class MarkLinkedDocCreateRequest
     {
         [Required]
-        public int LinkedDocId { get; set; }
+        public Int16 LinkedDocId { get; set; }
 
         [MaxLength(50)]
         public string Note { get; set; }
