+++ server/DocumentsKM/Dtos/MarkLinkedDoc/MarkLinkedDocUpdateRequest.cs
@@ -1,10 +1,11 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace DocumentsKM.Dtos
 {
     public class MarkLinkedDocUpdateRequest
     {
-        public int? LinkedDocId { get; set; }
+        public Int16? LinkedDocId { get; set; }
 
         [MaxLength(50)]
         public string Note { get; set; }
