+++ server/DocumentsKM/Dtos/MarkOperatingConditions/MarkOperatingConditionsRequest.cs
@@ -1,14 +1,16 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class MarkOperatingConditionsUpdateRequest
     {
         public float? SafetyCoeff { get; set; }
-        public int? EnvAggressivenessId { get; set; }
-        public int? Temperature { get; set; }
-        public int? OperatingAreaId { get; set; }
-        public int? GasGroupId { get; set; }
-        public int? ConstructionMaterialId { get; set; }
-        public int? PaintworkTypeId { get; set; }
-        public int? HighTensileBoltsTypeId { get; set; }
+        public Int16? EnvAggressivenessId { get; set; }
+        public Int16? Temperature { get; set; }
+        public Int16? OperatingAreaId { get; set; }
+        public Int16? GasGroupId { get; set; }
+        public Int16? ConstructionMaterialId { get; set; }
+        public Int16? PaintworkTypeId { get; set; }
+        public Int16? HighTensileBoltsTypeId { get; set; }
     }
 }
