+++ server/DocumentsKM/Dtos/MarkOperatingConditions/MarkOperatingConditionsResponse.cs
@@ -1,3 +1,4 @@
+using System;
 using DocumentsKM.Models;
 
 namespace DocumentsKM.Dtos
@@ -6,7 +7,7 @@ namespace DocumentsKM.Dtos
     {
         public float SafetyCoeff { get; set; }
         public EnvAggressiveness EnvAggressiveness { get; set; }
-        public int Temperature { get; set; }
+        public Int16 Temperature { get; set; }
         public OperatingArea OperatingArea { get; set; }
         public GasGroup GasGroup { get; set; }
         public ConstructionMaterial ConstructionMaterial { get; set; }
