+++ server/DocumentsKM/Dtos/StandardConstruction/StandardConstructionResponse.cs
@@ -1,10 +1,12 @@
+using System;
+
 namespace DocumentsKM.Dtos
 {
     public class StandardConstructionResponse
     {
-        public int Id { get; set; }
+        public Int16 Id { get; set; }
         public string Name { get; set; }
-        public int? Num { get; set; }
+        public Int16? Num { get; set; }
         public string Sheet { get; set; }
         public float Weight { get; set; }
     }
